Add files via upload
The word file is the description of the code and shows the procedure to follow. The excel files are the input matrix data for the 4 groups Client, Consultants, Contractors, and Freelancers. The file M1 is for the 1st level matrix and M2, M3, and M4 are respective files for the 2nd level division as shown in Figure5 of the word files. So the user can change the factors as per their requirements.